move test helpers to helpers file
